pyproject.toml
[tool.poetry]
name = "metrics-demo"
version = "0.1.0"
description = ""
authors = ["Amanda Walker <amanda@amandawalker.io>"]

[tool.poetry.dependencies]
python = "^3.7"
streamlit-nightly = "^1.1.1-alpha.20211104"
pandas = "*"
numpy = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
